run "react-native upgrade"
